+++ packages/jest-types/src/Config.ts
@@ -338,9 +338,11 @@ export type ProjectConfig = {
   cache: boolean;
   cacheDirectory: string;
   clearMocks: boolean;
+  collectCoverage: boolean;
   collectCoverageFrom: Array<string>;
   coverageDirectory: string;
   coveragePathIgnorePatterns: Array<string>;
+  coverageProvider: CoverageProvider;
   coverageReporters: CoverageReporters;
   cwd: string;
   dependencyExtractor?: string;